monitor/requirements.txt
# OpenTelemetry Core
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0

# Instrumentations (matching stable versions)
opentelemetry-instrumentation-django==0.48b0
opentelemetry-instrumentation-logging==0.48b0
opentelemetry-instrumentation-redis==0.48b0
opentelemetry-instrumentation-mysql==0.48b0
opentelemetry-instrumentation==0.48b0

# Semantic conventions
opentelemetry-semantic-conventions==0.48b0

# Django
Django>=5.0,<6.0
djangorestframework==3.14.0

# Utilities
python-dotenv==1.0.1
prometheus-client
python-json-logger
grpcio==1.64.1
redis>=4.0.0
mysqlclient==2.2.7
